% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the unique identification of a U.S. state}

\item{year}{relevant data year}
}
\description{
This function maps accidents in individual U.S. State in a given year
}
\details{
For a given year value, this function read the relevant data file
using the \code{\link{make_filename}} and \code{\link{fars_read}} functions.
It checks that the stae exists and that any accidents were reported that year in that state.
The function also removes erroneous longotude and lattitudes entries in the raw data
(\code{longitude>900} and \code{lattitude>90}) and uses the \code{\link{map}} package to
draw the relevant map and the \code{\link{graphics}} package to plot dots.
}
\examples{
\dontrun{
fars_map_state("10","2013")
}

}

