% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Data date range}
\usage{
fars_read_years(years)
}
\description{
This function returns the month and year of the data in a range of annual data files
}
\details{
@details This function iterates over a range of year values and uses the \code{\link{fars_read}} and \code{\link{make_filename}}
 to find and report the content of the MONTH and YEAR columns in each data file. The data files have to be in the same working directory.

@param years a vector of integer year values (YYYY)
 @inheritParams fars_read
 @inheritParams make_filename

@importFrom dyplr mutate select

@return A tipple of the MONTH and YEAR values for each data file in the \code{years} range
 @examples
 \dontrun{
 fars_read_years(c(2013:2015))
 }

@export
}

