% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary statistics}
\usage{
fars_summarize_years(years)
}
\value{
table of summary statistics
}
\description{
This function provides summary monthly statistics for each year in a range
}
\details{
This function uses the output from \code{\link{fars_read_years}}
to generate summary accident statistics by \code{YEAR} and \code{MONTH}.
}
\examples{
\dontrun{
fars_summarize_years(c(2013:2015))
}

}

