% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_data_companies_current}
\alias{infra_data_companies_current}
\title{Production and farm value of maple products in Canada}
\format{A data frame with columns:
\describe{
 \item{Year}{2000 to 2015}
 \item{Syrup}{Maple products expressed as syrup, total in thousands of gallons.}
 \item{CAD}{Gross value of maple products in thousands of Canadian dollars.}
 \item{Region}{Postal code abbreviation for territory or province.}
}}
\source{
EDHECinfra database, 27jan17
 maple products, annual. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
infra_data_companies_current
}
\description{
Production and farm value of maple products in Canada
}
\examples{
\dontrun{
 infra_data_companies_current
}
}
\keyword{datasets}

