% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Standard data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer year value (YYYY)}
}
\value{
a string representing a standard file name for a given year
}
\description{
Standard data file name
}
\details{
This function returns a standard name  for a given year for the source zip files
from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\examples{
\dontrun{
make_filename("accident_\%d.csv.bz2", 2013)
Creates a standard file name for the 2017 dataset
}

}

