% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{testing}
\alias{testing}
\title{Testing}
\usage{
testing()
}
\value{
test results from the testthat package
}
\description{
This function runs the tests in the tests/ directory. it has not argument
}
\examples{
\dontrun{
testing()
}

}

