% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read source data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string and, optionally a path, representing a CSV file name}
}
\value{
a dplyr dataframe
}
\description{
Read source data file
}
\details{
This function looks for a CSV file called \code{filename} and checks whether it exists or not,
if found it loads the data using \code{readr::read.csv} and converts it to a dyplr dataframe using \code{dyplr::tbl_df}.
If no data file with that name exists, the funtion returns an error.
}
\examples{
\dontrun{
fars_read("accident_2013.csv")}

}

